requirements.txt
fastapi==0.111.0
uvicorn==0.22.0
sqlalchemy==2.0.28
pydantic==2.6.4
pydantic-settings==2.2.1
alembic==1.7.3
aiomysql==0.2.0
passlib==1.7.4
python-jose==3.3.0
cachetools==4.2.4